add the test vec function fixup fps camera move error fixup Draw::view function error

+++ include/camera_base.h
@@ -3,7 +3,7 @@
 class CameraBase {
 public:
 	virtual ~CameraBase() = default;
-	virtual void key_callback(Window::WindowKey key, int delte_time) = 0;
+	virtual void key_callback(Window::WindowKey key, float delte_time) = 0;
 	virtual void mouse_callback(int x, int y) = 0;
 	virtual void scroll_callback(float offset) = 0;
 	virtual void frame_callback(int width, int height) = 0;
